ПР14/feature: Добавлена возможность регистрации и авторизации пользователя

+++ controllers/cards.js
@@ -1,44 +1,51 @@
 const Card = require('../models/card');
-const { BAD_REQUEST_CODE, NOT_FOUND_CODE, DEFAUTL_CODE } = require('../utils/constants');
+const NotFoundError = require('../errors/notFoundError');
+const BadRequestError = require('../errors/badRequestError');
+const AccessDeniedError = require('../errors/accessDeniedError');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(DEFAUTL_CODE).send({ message: 'Ошибка по умолчанию.' }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST_CODE).send({ message: 'Переданы некорректные данные при создании карточки.' });
+        next(new BadRequestError('Переданы некорректные данные при создании карточки.'));
       } else {
-        res.status(DEFAUTL_CODE).send({ message: 'Ошибка по умолчанию.' });
+        next(err);
       }
     });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        res.status(NOT_FOUND_CODE).send({ message: 'Карточка с указанным id не найдена' });
-      } else {
-        res.send({ data: card });
+        throw new NotFoundError('Карточка с указанным id не найдена');
       }
+      if (String(card.owner) !== req.user._id) {
+        throw new AccessDeniedError('Недостаточно прав');
+      }
+      Card.findByIdAndRemove(req.params.cardId)
+        .then((cardForDelete) => {
+          res.send({ data: cardForDelete });
+        });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(BAD_REQUEST_CODE).send({ message: 'Передан неккоректный _id для удаления карточки.' });
+        next(new BadRequestError('Передан неккоректный _id для удаления карточки'));
       } else {
-        res.status(DEFAUTL_CODE).send({ message: 'Ошибка по умолчанию.' });
+        next(err);
       }
     });
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -46,21 +53,21 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        res.status(NOT_FOUND_CODE).send({ message: 'Передан несуществующий _id карточки.' });
+        throw new NotFoundError('Карточка с указанным id не найдена');
       } else {
         res.send({ data: card.likes });
       }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(BAD_REQUEST_CODE).send({ message: 'Переданы некорректные данные для постановки/снятии лайка.' });
+        next(new BadRequestError('Переданы некорректные данные для постановки лайка'));
       } else {
-        res.status(DEFAUTL_CODE).send({ message: 'Ошибка по умолчанию.' });
+        next(err);
       }
     });
 };
 
-module.exports.deleteLikeCard = (req, res) => {
+module.exports.deleteLikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -68,16 +75,16 @@ module.exports.deleteLikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        res.status(NOT_FOUND_CODE).send({ message: 'Передан несуществующий _id карточки.' });
+        throw new NotFoundError('Карточка с указанным id не найдена');
       } else {
         res.send({ data: card.likes });
       }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(BAD_REQUEST_CODE).send({ message: 'Переданы некорректные данные для постановки/снятии лайка.' });
+        next(new BadRequestError('Переданы некорректные данные для снятия лайка'));
       } else {
-        res.status(DEFAUTL_CODE).send({ message: 'Ошибка по умолчанию.' });
+        next(err);
       }
     });
 };
